+++ Makefile
@@ -1,4 +1,4 @@
 .DEFAULT_GOAL := test
 
 test:
-	pylint tap_sftp -d missing-docstring,fixme,duplicate-code,line-too-long
+	pylint tap_nicesftp -d missing-docstring,fixme,duplicate-code,line-too-long
