+++ tap_nicesftp/singer_encodings/compression.py
