+++ tests/tox_tests/test_med_00_compression.py
@@ -1,4 +1,4 @@
-from tap_sftp.singer_encodings.compression import infer
+from tap_nicesftp.singer_encodings.compression import infer
 from tests.configuration.fixtures import get_sample_file_path
