+++ tests/tox_tests/test_med_00_discover.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timezone
 from unittest.mock import patch
 
-from tap_sftp.discover import discover_streams
+from tap_nicesftp.discover import discover_streams
 from tests.configuration.fixtures import get_sample_file_path, get_table_spec
 
 expected_schema = {
