+++ tests/tox_tests/test_med_00_sync.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timezone
 from unittest.mock import patch
 
-from tap_sftp.sync import sync_file, sync_stream
+from tap_nicesftp.sync import sync_file, sync_stream
 from tests.configuration.fixtures import (get_catalog, get_sample_file_path,
                                           get_table_spec, sftp_client)
