+++ tox.ini
@@ -17,7 +17,7 @@ whitelist_externals =
     pycodestyle
     coverage
 commands =
-    pycodestyle tap_sftp/
+    pycodestyle tap_nicesftp/
     coverage run -m pytest {posargs}
     coverage html -d tests/codecoverage
 [pycodestyle]
@@ -30,7 +30,7 @@ testpaths =
     tests/tox_tests
 [coverage:run]
 source=
-    tap_sftp/
+    tap_nicesftp/
 omit=
     */.tox/*
     */test/*
